further refactoring. Could also finally expose a default export for the whole library. Now it will work as drop in replacement for bugsnag-js
